getStringNoSpecials(...) has been moved to anautilities.py to reduce import mismash discussed in #68
